+++ src/detail/esp_detail/mimetable.h
@@ -13,7 +13,7 @@
   @file       Esp8266WebServer.h
   @author     Ivan Grokhotkov
 
-  Version: 2.2.2
+  Version: 2.2.3
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -29,6 +29,7 @@
   2.2.0   K Hoang      05/05/2022 Add support to custom SPI for Teensy, Mbed RP2040, Portenta_H7, etc.
   2.2.1   K Hoang      25/08/2022 Auto-select SPI SS/CS pin according to board package
   2.2.2   K Hoang      06/09/2022 Slow SPI clock for old W5100 shield or SAMD Zero. Improve support for SAMD21
+  2.2.3   K Hoang      17/09/2022 Add support to AVR Dx (AVR128Dx, AVR64Dx, AVR32Dx, etc.) using DxCore
  *************************************************************************************************************************************/
 
 #pragma once
@@ -36,7 +37,7 @@
 #ifndef __ESP_MIMETABLE_H__
 #define __ESP_MIMETABLE_H__
 
-#if (ESP32 || ESP8266)
+#if (defined(ESP32) || defined(ESP8266))
 
 #include "WString.h"
 
@@ -84,5 +85,5 @@ namespace mime_esp
   String getContentType(const String& path);
 }
 
-#endif    // #if (ESP32 || ESP8266)
+#endif    // #if (defined(ESP32) || defined(ESP8266))
 #endif    // #ifndef __ESP_MIMETABLE_H__